extract SummaryMetrics component
